- Adicao das informacoes na tab de conta - Adicao de box de imovel informativo - Ajuste no DataManipulator para a execucao correta do calculo de conta

+++ src/com/IS/ListaImoveis.java
@@ -24,7 +24,7 @@ import business.ControladorRota;
 public class ListaImoveis extends ListActivity {
 	
 	MySimpleArrayAdapter enderecoList;
-	public static int tamanhoListaImoveis;
+	public static long tamanhoListaImoveis;
 //	ArrayList<String> listStatusImoveis;
 
     /** Called when the activity is first created. */
